# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API SIAKAD GERBANG",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/reference/education-levels": {
            "get": {
                "tags": [
                    "Reference/EducationLevel"
                ],
                "summary": "List all EducationLevel",
                "operationId": "9e835793477699121e61c21d3548961f",
                "responses": {
                    "200": {
                        "description": "List of EducationLevel",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EducationLevelListResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Reference/EducationLevel"
                ],
                "summary": "Create a new EducationLevel",
                "operationId": "97b8c4c7b641f9537e8812e963c33f57",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "object",
                                        "example": "example name"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "example value"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "EducationLevel created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EducationLevelResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                }
            }
        },
        "/api/reference/education-levels/{id}": {
            "get": {
                "tags": [
                    "Reference/EducationLevel"
                ],
                "summary": "Get EducationLevel detail by ID",
                "operationId": "04c4ea397f3e51f12e43409db6a329b1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "EducationLevel ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "EducationLevel detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EducationLevelResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "EducationLevel not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Reference/EducationLevel"
                ],
                "summary": "Update an existing EducationLevel",
                "operationId": "195b490ceebaf91830f2a6c1e0297920",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "EducationLevel ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "object",
                                        "example": "example name"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "example value"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "EducationLevel updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EducationLevelResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "EducationLevel not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Reference/EducationLevel"
                ],
                "summary": "Delete a EducationLevel by ID",
                "operationId": "b68338d5dcf79567d2aa417253a725e8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "EducationLevel ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "EducationLevel deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": "success",
                                    "message": "Data deleted",
                                    "data": true
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "EducationLevel not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "EducationLevel": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "properties": {
                            "en": {
                                "type": "string",
                                "example": "Bachelor"
                            },
                            "id": {
                                "type": "string",
                                "example": "Sarjana"
                            }
                        },
                        "type": "object"
                    },
                    "code": {
                        "type": "string",
                        "example": "S1"
                    }
                },
                "type": "object"
            },
            "EducationLevelResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": "Education Level detail"
                    },
                    "data": {
                        "$ref": "#/components/schemas/EducationLevel"
                    }
                },
                "type": "object"
            },
            "EducationLevelListResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": "List of education levels"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/EducationLevel"
                        }
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}